[FEAT] 컨트롤러 토큰 관련 전반적 수정

+++ karim/src/main/java/com/karim/karim/controller/PlanController.java
@@ -5,7 +5,9 @@ import com.karim.karim.dto.PlaceListDto;
 import com.karim.karim.dto.PlanDto;
 import com.karim.karim.service.PlaceService;
 import com.karim.karim.service.PlanService;
+import com.karim.karim.util.JWTUtil;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,10 +22,12 @@ public class PlanController {
 
     private final PlanService planService;
     private final PlaceService placeService;
+    private final JWTUtil jwtUtil;
 
-    public PlanController(PlanService planService, PlaceService placeService) {
+    public PlanController(PlanService planService, PlaceService placeService, JWTUtil jwtUtil) {
         this.planService = planService;
         this.placeService = placeService;
+        this.jwtUtil = jwtUtil;
     }
 
     @Operation(summary = "전체 계획 조회", description = "사용자가 등록한 모든 여행 계획을 조회합니다.")
@@ -35,8 +39,14 @@ public class PlanController {
 
     @Operation(summary = "계획 생성", description = "새로운 여행 계획을 등록합니다.")
     @PostMapping()
-    public ResponseEntity<String> save(@RequestBody PlanDto planDto) {
+    public ResponseEntity<String> save(
+            @RequestBody PlanDto planDto,
+            @Parameter(hidden = true) @RequestHeader(value = "Authorization", required = false) String accessToken) {
+
+        jwtUtil.validateAccessToken(accessToken, planDto.getUserId());
+
         int result = planService.savePlan(planDto);
+
         if (result > 0) {
             return ResponseEntity.ok("계획 생성 성공");
         } else {
@@ -46,8 +56,14 @@ public class PlanController {
 
     @Operation(summary = "계획 수정", description = "기존 여행 계획을 수정합니다.")
     @PatchMapping()
-    public ResponseEntity<String> modify(@RequestBody PlanDto planDto) {
+    public ResponseEntity<String> modify(
+            @RequestBody PlanDto planDto,
+            @Parameter(hidden = true) @RequestHeader(value = "Authorization", required = false) String accessToken) {
+
+        jwtUtil.validateAccessToken(accessToken, planDto.getUserId());
+
         int result = planService.modifyPlan(planDto);
+
         if (result > 0) {
             return ResponseEntity.ok("계획 수정 성공");
         } else {
@@ -56,9 +72,15 @@ public class PlanController {
     }
 
     @Operation(summary = "계획 삭제", description = "특정 ID의 여행 계획을 삭제합니다.")
-    @DeleteMapping("/{userId}/{id}")
-    public ResponseEntity<String> delete(@PathVariable int id,@PathVariable Long userId) {
-        int result = planService.deletePlan(id, userId);
+    @DeleteMapping("/{id}")
+    public ResponseEntity<String> delete(
+            @PathVariable int id,
+            @Parameter(hidden = true) @RequestHeader(value = "Authorization", required = false) String accessToken) {
+
+        jwtUtil.validateAccessToken(accessToken, planService.findById(id).getUserId());
+
+        int result = planService.deletePlan(id);
+
         if (result > 0) {
             return ResponseEntity.ok("계획 삭제 성공");
         } else {
@@ -67,9 +89,10 @@ public class PlanController {
     }
 
     @Operation(summary = "계획 상세 조회", description = "특정 여행 계획과 해당 계획의 장소 목록을 조회합니다.")
-    @GetMapping("/detail/{userId}/{id}")
-    public ResponseEntity<Map<String, Object>> getPlanDetails(@PathVariable("id") int id, @PathVariable("userId") Long userId) {
-        PlanDto plan = planService.findByIdAndUserId(id, userId);
+    @GetMapping("/detail/{id}")
+    public ResponseEntity<Map<String, Object>> getPlanDetails(@PathVariable("id") int id) {
+
+        PlanDto plan = planService.findById(id);
         List<PlaceDto> places = placeService.findByPlanId(id);
         Map<String, Object> response = new HashMap<>();
         response.put("plan", plan);
@@ -80,9 +103,17 @@ public class PlanController {
 
     @Operation(summary = "계획 내 장소 추가", description = "특정 계획에 새로운 장소를 추가합니다.")
     @PostMapping("/detail/{id}")
-    public ResponseEntity<String> addPlaceToPlan(@PathVariable("id")  int id, @RequestBody PlaceDto placeDto) {
+    public ResponseEntity<String> addPlaceToPlan(
+            @PathVariable("id")  int id,
+            @RequestBody PlaceDto placeDto,
+            @Parameter(hidden = true) @RequestHeader(value = "Authorization", required = false) String accessToken) {
+
+        jwtUtil.validateAccessToken(accessToken, planService.findById(id).getUserId());
+
         placeDto.setPlanId(id);
+
         int result = placeService.save(placeDto);
+
         if (result > 0) {
             return ResponseEntity.ok("장소 추가 성공");
         } else {
@@ -92,8 +123,15 @@ public class PlanController {
 
     @Operation(summary = "계획 내 장소 수정", description = "특정 계획 내 장소를 수정합니다.")
     @PutMapping("/detail/{id}")
-    public ResponseEntity<String> modifyPlaceInPlan(@PathVariable("id") int id, @RequestBody PlaceListDto placeListDto) {
+    public ResponseEntity<String> modifyPlaceInPlan(
+            @PathVariable("id") int id,
+            @RequestBody PlaceListDto placeListDto,
+            @Parameter(hidden = true) @RequestHeader(value = "Authorization", required = false) String accessToken) {
+
+        jwtUtil.validateAccessToken(accessToken, planService.findById(id).getUserId());
+
         int result = placeService.modify(placeListDto.getData());
+
         if (result > 0) {
             return ResponseEntity.ok("장소 수정 성공");
         } else {
@@ -103,12 +141,19 @@ public class PlanController {
 
     @Operation(summary = "계획 내 장소 삭제", description = "특정 계획에서 장소를 삭제합니다.")
     @DeleteMapping("/detail/{id}/{placeId}")
-    public ResponseEntity<String> deletePlaceFromPlan(@PathVariable("id") int id, @PathVariable("placeId") int placeId) {
+    public ResponseEntity<String> deletePlaceFromPlan(
+            @PathVariable("id") int id,
+            @PathVariable("placeId") int placeId,
+            @Parameter(hidden = true) @RequestHeader(value = "Authorization", required = false) String accessToken) {
+
+        jwtUtil.validateAccessToken(accessToken, planService.findById(id).getUserId());
+
         PlaceDto placeDto = new PlaceDto();
         placeDto.setId(placeId);
         placeDto.setPlanId(id);
 
         int result = placeService.delete(placeDto);
+
         if (result > 0) {
             return ResponseEntity.ok("장소 삭제 성공");
         } else {
